refactor project state into Body from AllProjects * Body passes projects as props to AllProjects * probably need to refactor how Body makes the db call; this is messy right now
