+++ lib/components/AllProjects.jsx
@@ -1,25 +1,9 @@
 import React from 'react';
-import 'firebase/app'
-import 'firebase/database'
-import firebase from 'firebase';
 
 export default class AllProjects extends React.Component{
-  constructor() {
-    super();
-    this.state = { projects: [] }
-    let app = firebase.initializeApp({ databaseURL: "https://project-manager-18532.firebaseio.com"});
-    let db = firebase.database();
-    this.ref = db.ref("projects")
-  }
-
-  componentDidMount() {
-    this.ref.on("value", (dataSnapshot) => {
-      this.setState({projects: dataSnapshot.val()})
-    })
-  }
 
   render() {
-    let projectObject = this.state.projects
+    let projectObject = this.props.projects
     let projects = [];
     for (var key in projectObject) {
       if(projectObject.hasOwnProperty(key)) {
@@ -31,7 +15,6 @@ export default class AllProjects extends React.Component{
         projects.push(html)
       }
     }
-    console.log("result:", projects)
 
     return (
       <div className="all-projects">
